Refactoring bootstrap files, import and utilities
